[fixed] potential read out of bounds when decrypting incoming incorrect packets

+++ MMOEngine/src/system/lang/Integer.cpp
@@ -35,7 +35,7 @@ uint32 Integer::hexvalueOf(const String& str) {
 		mul *= 16;
 	}
 
-	return val;
+	return (uint32) (val);
 }
 
 bool Integer::toString(String& str) {
